# docs/api/swagger.yaml
definitions:
  APIErrorResponse:
    properties:
      code:
        type: integer
      details: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  internal_modules_project_handler.DeleteResponse:
    properties:
      status:
        type: string
    type: object
...
  internal_modules_project_handler.GooglePlayCredentialsCreateRequest:
    properties:
      credentials:
        example: base64-encoded service account JSON
        type: string
      name:
        example: My Service Account
        type: string
    type: object
  internal_modules_project_handler.GooglePlayCredentialsDTO:
    properties:
      created_at:
        type: string
      id:
        example: 1
        type: integer
      name:
        example: My Service Account
        type: string
      updated_at:
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  internal_modules_project_handler.GooglePlayCredentialsListResponse:
    properties:
      google_play_credentials:
        items:
          $ref: '#/definitions/internal_modules_project_handler.GooglePlayCredentialsDTO'
        type: array
    type: object
  internal_modules_project_handler.GooglePlayCredentialsResponse:
    properties:
      google_play_credentials:
        $ref: '#/definitions/internal_modules_project_handler.GooglePlayCredentialsDTO'
    type: object
...
info:
  contact:
    email: support@flotio.dev
    name: Flotio Team
  description: Flotio Core API — authentication, environment/keystore/Google Play
    credentials assets, projects, build pipeline, releases and GitHub integration
    for the Flotio mobile CI platform.
  license:
    name: Proprietary
  title: Flotio Core API
  version: 1.0.0
paths:
...
  /google-play-credentials:
    get:
      consumes:
      - application/json
      description: Get all Google Play distribution credentials owned by the authenticated
        user
      operationId: GooglePlayCredentialsGetHandler
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.GooglePlayCredentialsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user Google Play credentials
      tags:
      - google-play-credentials
    post:
      consumes:
      - application/json
      description: Upload new Google Play distribution credentials for the authenticated
        user
      operationId: GooglePlayCredentialsPostHandler
      parameters:
      - description: Google Play credentials data
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_modules_project_handler.GooglePlayCredentialsCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_modules_project_handler.GooglePlayCredentialsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Google Play credentials
      tags:
      - google-play-credentials
  /google-play-credentials/{credentialsId}:
    delete:
      consumes:
      - application/json
      description: Delete a specific Google Play credentials entry by ID
      operationId: GooglePlayCredentialsDeleteHandler
      parameters:
      - description: Credentials ID
        format: int64
        in: path
        name: credentialsId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.DeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Google Play credentials
      tags:
      - google-play-credentials
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"